@@ -1,6 +1,6 @@
 ﻿using System.Globalization;
 using Microsoft.Extensions.Logging;
-using Tauron.Akka;
+using Tauron.TAkka;
 using Tauron.Application.Settings;
 
 namespace Akka.MGIHelper.Core.Configuration
